+++ src/CrossValidation/Reports/ContingencyTable.php
@@ -9,7 +9,8 @@ use InvalidArgumentException;
  * Contingency Table
  *
  * A Contingency Table is used to display the frequency distribution of class
- * labels among a clustering of samples.
+ * labels among a clustering of samples. It is similar to a Confusion Matrix
+ * but uses the labels to establish a ground truth for a clustering instead.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -17,33 +18,6 @@ use InvalidArgumentException;
  */
 class ContingencyTable implements Report
 {
-    /**
-     * The classes to compare in the table.
-     *
-     * @var array|null
-     */
-    protected $classes;
-
-    /**
-     * @param array|null $classes
-     * @throws \InvalidArgumentException
-     */
-    public function __construct(?array $classes = null)
-    {
-        if (is_array($classes)) {
-            $classes = array_unique($classes);
-
-            foreach ($classes as $class) {
-                if (!is_string($class) and !is_int($class)) {
-                    throw new InvalidArgumentException('Class type must be a'
-                        . ' string or integer, ' . gettype($class) . ' found.');
-                }
-            }
-        }
-
-        $this->classes = $classes;
-    }
-
     /**
      * The estimator types that this report is compatible with.
      *
@@ -71,18 +45,13 @@ class ContingencyTable implements Report
                 . ' must equal the number of predictions.');
         }
 
-        $classes = $this->classes ?? array_unique($labels);
-
+        $classes = array_unique($labels);
         $clusters = array_unique($predictions);
 
         $table = array_fill_keys($clusters, array_fill_keys($classes, 0));
 
-        $included = array_flip($classes);
-
-        foreach ($labels as $i => $label) {
-            if (isset($included[$label])) {
-                $table[$predictions[$i]][$label]++;
-            }
+        foreach ($labels as $i => $class) {
+            $table[$predictions[$i]][$class]++;
         }
 
         return $table;