+++ src/NeuralNet/ActivationFunctions/ReLU.php
@@ -7,12 +7,14 @@ use Rubix\Tensor\Matrix;
 /**
  * ReLU
  *
- * Rectified Linear Units output only the positive part of its inputs and are
- * analogous to a half-wave rectifiers in electrical engineering.
+ * A Thresholded ReLU (Rectified Linear Unit) only outputs the signal above
+ * a user-defined threshold parameter.
  *
  * References:
- * [1] V. Nair et al. (2011). Rectified Linear Units Improve Restricted
- * Boltzmann Machines.
+ * [1] A. L. Maas et al. (2013). Rectifier Nonlinearities Improve Neural
+ * Network Acoustic Models.
+ * [2] K. Konda et al. (2015). Zero-bias Autoencoders and the Benefits of
+ * Co-adapting Features.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -21,6 +23,22 @@ use Rubix\Tensor\Matrix;
 class ReLU implements ActivationFunction
 {
     /**
+     * The input value necessary to trigger an activation.
+     *
+     * @var float
+     */
+    protected $threshold;
+
+    /**
+     * @param float $threshold
+     * @throws \InvalidArgumentException
+     */
+    public function __construct(float $threshold = 0.)
+    {
+        $this->threshold = $threshold;
+    }
+
+    /**
      * Return a tuple of the min and max output value for this activation
      * function.
      *
@@ -28,7 +46,7 @@ class ReLU implements ActivationFunction
      */
     public function range() : array
     {
-        return [0., INF];
+        return [min(0., $this->threshold), INF];
     }
 
     /**
@@ -60,7 +78,7 @@ class ReLU implements ActivationFunction
      */
     public function _compute(float $z) : float
     {
-        return $z > 0. ? $z : 0.;
+        return $z > $this->threshold ? $z : 0.;
     }
 
     /**
@@ -69,6 +87,6 @@ class ReLU implements ActivationFunction
      */
     public function _differentiate(float $computed) : float
     {
-        return $computed > 0. ? 1. : 0.;
+        return $computed > $this->threshold ? 1. : 0.;
     }
 }