+++ src/NeuralNet/ActivationFunctions/SELU.php
@@ -3,13 +3,14 @@
 namespace Rubix\ML\NeuralNet\ActivationFunctions;
 
 use Rubix\Tensor\Matrix;
-use InvalidArgumentException;
 
 /**
  * SELU
  *
  * Scaled Exponential Linear Unit is a self-normalizing activation function
- * based on the ELU activation function.
+ * based on the ELU activation function. Neuronal activations of SELU networks
+ * automatically converge toward zero mean and unit variance, unlike explicitly
+ * normalized networks such as those with [Batch Norm](#batch-norm).
  *
  * References:
  * [1] G. Klambauer et al. (2017). Self-Normalizing Neural Networks.
@@ -20,52 +21,10 @@ use InvalidArgumentException;
  */
 class SELU implements ActivationFunction
 {
-    protected const ALPHA = 1.6732632423543772848170429916717;
-    protected const SCALE = 1.0507009873554804934193349852946;
+    public const ALPHA = 1.6732632423543772848170429916717;
+    public const SCALE = 1.0507009873554804934193349852946;
 
-    /**
-     * At which negative value the SELU will saturate. i.e. alpha = 1.means
-     * that the leakage will never be more than -1.0.
-     *
-     * @var float
-     */
-    protected $alpha;
-
-    /**
-     * The scaling factor.
-     *
-     * @var float
-     */
-    protected $scale;
-
-    /**
-     * The exponential leakage coefficient.
-     *
-     * @var float
-     */
-    protected $beta;
-
-    /**
-     * @param float $alpha
-     * @param float $scale
-     * @throws \InvalidArgumentException
-     */
-    public function __construct(float $alpha = self::ALPHA, float $scale = self::SCALE)
-    {
-        if ($alpha < 0.) {
-            throw new InvalidArgumentException('Alpha parameter must be'
-                . " greater than 0, $alpha given.");
-        }
-
-        if ($scale < 1.) {
-            throw new InvalidArgumentException('Scale must be greater than'
-                . " 1, $scale given.");
-        }
-
-        $this->scale = $scale;
-        $this->alpha = $alpha;
-        $this->beta = $scale * $alpha;
-    }
+    protected const BETA = self::SCALE * self::ALPHA;
 
     /**
      * Return a tuple of the min and max output value for this activation
@@ -75,7 +34,7 @@ class SELU implements ActivationFunction
      */
     public function range() : array
     {
-        return [-$this->beta, INF];
+        return [-self::BETA, INF];
     }
 
     /**
@@ -107,7 +66,7 @@ class SELU implements ActivationFunction
      */
     public function _compute(float $z) : float
     {
-        return $z > 0. ? $this->scale * $z : $this->beta * (exp($z) - 1.);
+        return $z > 0. ? self::SCALE * $z : self::BETA * (exp($z) - 1.);
     }
 
     /**
@@ -116,6 +75,6 @@ class SELU implements ActivationFunction
      */
     public function _differentiate(float $computed) : float
     {
-        return $computed > 0. ? $this->scale : $computed + 1.;
+        return $computed > 0. ? self::SCALE : self::SCALE * ($computed + self::ALPHA);
     }
 }