+++ src/Parallel.php
@@ -5,16 +5,16 @@ namespace Rubix\ML;
 interface Parallel
 {
     /**
-     * Set the maximum number of processes to run in parallel.
-     *
-     * @param int $n
-     */
-    public function setWorkers(int $n) : void;
-
-    /**
      * Return the maximum number of workers.
      *
      * @return int
      */
     public function workers() : int;
+
+    /**
+     * Set the maximum number of processes to run in parallel.
+     *
+     * @param int $n
+     */
+    public function setWorkers(int $n) : void;
 }